Read the SSH keys used to encrypt/decrypt from configured paths
Previously, the path to the private key was found by looking at
each key stored in the SSH agent in turn. It happens that this
information is no always available however: on recent MacBook Pro
computers, only the public key is included in `ssh-add -L` output,
not the path to the private key.

I added parameters in `config.sh` for the locations of the SSH keys
used to encrypt/decrypt on the local machine.

Related issue: #2